last touches on blueprint system before merge(code is unclean ik)
