+++ src/source.rs
@@ -104,10 +104,15 @@ pub fn type_mangle(name: String, types: Vec<ConstType>) -> String {
     let mut mangle = String::new();
     mangle.push_str(name.as_str());
 
+    // if we only have no args, later we should analyze blueprints with only 1 possible instance
+    if types.len() == 0 {
+        mangle.push_str("_void");
+    }
     for type_n in types {
         mangle.push('_');
         mangle.push_str(type_n.as_str());
     }
+
     return mangle;
 }
 impl Enviroment {
@@ -117,19 +122,6 @@ impl Enviroment {
             current: ConstType::Void,
             parent,
             blueprints: Vec::new(),
-        }
-    }
-
-    pub fn blueprints(&mut self, blueprints: Vec<Blueprint>) {
-        self.blueprints = blueprints.clone();
-        for blueprint in blueprints {
-            self.add(
-                &blueprint.name,
-                ConstType::Blueprint {
-                    argc: blueprint.args.len() as u32,
-                    name: blueprint.name.clone(),
-                },
-            );
         }
     }
